Engin graphique extrait dans un projet separer qui est inclue comme dependence
