+++ CMakeLists.txt
@@ -85,3 +85,8 @@ add_subdirectory( linesCore )
 add_subdirectory( linesNetwork )
 add_subdirectory( linesDesktopLib )
 add_subdirectory( linesDesktop )
+
+
+include_directories( GraphicsEngine )
+add_subdirectory( GraphicsEngine )
+target_link_libraries( GraphicsEngine )
