+++ GraphicsEngine/GraphicsEngine.h
@@ -1,0 +1,7 @@
+#ifndef GraphicsEngine_H
+#define GraphicsEngine_H
+
+#include "include/IGraphicsEngine.h"
+#include "include/OpenGLGraphicsEngine.h"
+
+#endif // GraphicsEngine_H
