+++ linesCore/drwDrawableTexture.cpp
@@ -1,6 +1,4 @@
 #include "drwDrawableTexture.h"
-#include "IncludeGLEW.h"
-#include "IncludeGl.h"
 #include <assert.h>
 
 drwDrawableTexture::drwDrawableTexture()
